+++ core/src/main/java/org/keycloak/representations/idm/RealmRepresentation.java
@@ -50,6 +50,7 @@ public class RealmRepresentation {
     protected Boolean revokeRefreshToken;
     protected Integer refreshTokenMaxReuse;
     protected Integer accessTokenLifespan;
+    protected Integer idTokenLifespan;
     protected Integer accessTokenLifespanForImplicitFlow;
     protected Integer ssoSessionIdleTimeout;
     protected Integer ssoSessionMaxLifespan;
@@ -330,6 +331,14 @@ public class RealmRepresentation {
 
     public void setAccessTokenLifespan(Integer accessTokenLifespan) {
         this.accessTokenLifespan = accessTokenLifespan;
+    }
+
+    public Integer getIdTokenLifespan() {
+        return idTokenLifespan;
+    }
+
+    public void setIdTokenLifespan(Integer idTokenLifespan) {
+        this.idTokenLifespan = idTokenLifespan;
     }
 
     public Integer getAccessTokenLifespanForImplicitFlow() {