+++ services/src/main/java/org/keycloak/protocol/oidc/OIDCConfigAttributes.java
@@ -52,6 +52,7 @@ public final class OIDCConfigAttributes {
     public static final String ACCESS_TOKEN_SIGNED_RESPONSE_ALG = "access.token.signed.response.alg";
 
     public static final String ACCESS_TOKEN_LIFESPAN = "access.token.lifespan";
+    public static final String ID_TOKEN_LIFESPAN = "id.token.lifespan";
     public static final String CLIENT_SESSION_IDLE_TIMEOUT = "client.session.idle.timeout";
     public static final String CLIENT_SESSION_MAX_LIFESPAN = "client.session.max.lifespan";
     public static final String CLIENT_OFFLINE_SESSION_IDLE_TIMEOUT = "client.offline.session.idle.timeout";