+++ CHANGELOG.md
@@ -9,6 +9,8 @@ For Keycloak upstream changelog please see https://www.keycloak.org/docs/latest/
 Full Keycloak upstream jira issue can be shown if filtered by Fix version. 
 
 ## [v16.1.0-rc1.0.8] - 2022-03-28
+### Added
+- Id token lifespan [RCIAM-930](https://jira.argo.grnet.gr/browse/RCIAM-930)
 ### Changed
 - Support for configuring claims supported in Keycloak OP metadata [RCIAM-899](https://jira.argo.grnet.gr/browse/RCIAM-899)
 ### Fixed